Completely rewrote hosting.wizard.inc.
Started out as integrating documentation more cleanly, but the next thing I realized i had
come across a very clean and simple recipe for a wizard using only hook_menu and formsapi.

Split up requirements documentation in previous commits to allow for re-use in the new wizard,
and write a large amount of documentation additionally.

New wizard inherits a lot of the existing forms as well, and has additional validation. Changes
are mostly contained in hosting.wizard.inc ( apart from a couple of small changes in hosting.module
and a new css file)

New form is functionaly in line with the previous one, but is a whole lot cleaner. Has a few more extra pages, including some place holders that need to be properly finished before a release.