1222208 fix cron to not block on broken sites
we explode the listing function into a 'while' queue and ensure we count
sites only if their timestamp has been updated, otherwise there's the
potential (if the number of broken sites is higher than the number of
parallel cron jobs allowed) of always blocking over the same sites.

also, do not completely break out of the loop but continue to the next
iteration so that one failure doesn't crash the whole queue.