handle missing context information better
we produce a human-readable error message that is easier to track now
instead of a "Trying to get property of non-object" error.

This can happen when there's an inconsistency in the frontend
database, which can unfortunately happen (for example when you created
a platform without a webserver because of missing error-checking or
erroneous manual data entry).

This is to help diagnosing #1896184