move eldir back into its own module
this decouples the releases of core from aegir, the same way we do
(e.g.) with hosting_platform_pathauto.

the logic here is detailed in #1468874, but basically, eldir is
updated less frequently than other modules, so it's less likely to
have frequent updates. also, splitting it out will allow the theme to
be used by non-aegir installations, and facilitate the D7 port.