1968226 - manage IPs individually
the ip_addresses datastructure is now indexed as within the SQL
database, with the serial being the key to the array.

this allows us to individually add/change/remove entries in the IP
list of a server.

this seems to work very well in my manual tests, but it could use a
bit more AJAXy polish and further testing.