move hosting back to its own module
this decouples the releases of core from aegir, the same way we do
with eldir

we also change the makefiles so that the makefile found by drupal.org
will use the tagged releases while our local makefile will use the git
releases. the trick here is that the hostmaster.make file is first
read by drush make everywhere *except* on drupal.org.